reworked our chat element to be at the parent (game) level
