--- client/src/components/chat/index.js
@@ -1,3 +1,0 @@
-import Chat from './chat';
-
-export default Chat;
