--- client/src/components/gameBoard/index.js
@@ -1,3 +1,0 @@
-import GameBoard from './gameBoard'
-
-export default GameBoard;
